Update the data type of squarefootPerCarton from "double" to "BigDecimal" and also restrict the precision and scale. Update the data type of squarefoot_per_carton from "float(4)" to "decimal(10,2) to ensure data mapping: decimal -> BigDecimal

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -73,7 +73,7 @@ create table plank_sizes
     width_in_inch   varchar(255) not null,
     length      varchar(255) not null,
     thickness_in_inch   varchar(255) not null,
-    squarefoot_per_carton   float not null,
+    squarefoot_per_carton   decimal(10,2) not null,
     create_time  timestamp default null,
     update_time timestamp default null,
     create_user_id int8,
